Fix log spam if a client without the rerun protocol header tries to connect
