+++ crates/re_viewer/src/app.rs
@@ -983,15 +983,15 @@ impl App {
                             format!("Data source {} has left unexpectedly: {err}", msg.source);
 
                         #[cfg(not(target_arch = "wasm32"))]
-                        #[cfg(feature = "server")]
                         if err
                             .downcast_ref::<re_sdk_comms::ConnectionError>()
                             .is_some_and(|e| {
                                 matches!(e, re_sdk_comms::ConnectionError::UnknownClient)
                             })
                         {
-                            // An unknown client that probably stumbled onto the wrong port.
-                            // Don't log as an error (https://github.com/rerun-io/rerun/issues/5883).
+                            // This can happen if a client tried to connect but didn't send the `re_sdk_comms::PROTOCOL_HEADER`.
+                            // Likely an unknown client stumbled onto the wrong port - don't log as an error.
+                            // (for more information see https://github.com/rerun-io/rerun/issues/5883).
                             re_log::debug!("{log_msg}");
                             continue;
                         }
